[Fix] Handle submitting while fetching
The best way to handle a submit while fetching is actually to not
handle it :D

A submit can be checked for completion, and if successful, one simply
has to update the list of rooms, either locally or through an actual
fetch request.

If a user loads the page (fetching the rooms) and create a new room that
is not persisted before the fetch request is complete, then what happens
is:

[page-load] => fetch => [UI-updated]
       [user-submit] => submit => fetch => [UI-updated]

It's two fetch requests, but that's quite lightweight (because the list
of rooms is not that huge, especially since we moved the messages out),
and that way the UI is always consistent with the persisted data.